+++ molviewspec/app/api/examples.py
@@ -19,6 +19,7 @@ from molviewspec.nodes import (
     ContinuousPalette,
     DiscretePalette,
     GlobalMetadata,
+    LabelAttachmentT,
     PrimitiveComponentExpressions,
     RepresentationTypeT,
     Snapshot,
@@ -1031,6 +1032,36 @@ async def volume_map_example() -> MVSResponse:
         volume.representation(type="isosurface", relative_isovalue=1, show_wireframe=True)
         .color(color="blue")
         .opacity(opacity=0.66)
+    )
+
+    return JSONResponse(builder.get_state().to_dict())
+
+
+@router.get("/volume/slices")
+async def volume_map_example() -> MVSResponse:
+    """
+    Renders a volume in MAP format
+    """
+
+    builder = create_builder()
+
+    download = builder.download(url="https://www.ebi.ac.uk/pdbe/entry-files/1tqn.ccp4")
+    volume = download.parse(format="map").volume()
+
+    (
+        volume.representation(type="grid_slice", dimension="x", relative_index=0.5, relative_isovalue=1)
+        .color(color="red")
+        .opacity(opacity=0.75)
+    )
+    (
+        volume.representation(type="grid_slice", dimension="y", relative_index=0.5, relative_isovalue=1)
+        .color(color="green")
+        .opacity(opacity=0.75)
+    )
+    (
+        volume.representation(type="grid_slice", dimension="z", relative_index=0.5, relative_isovalue=1)
+        .color(color="blue")
+        .opacity(opacity=0.75)
     )
 
     return JSONResponse(builder.get_state().to_dict())
@@ -2304,6 +2335,71 @@ async def testing_instance_id_annot_selector() -> MVSResponse:
     return JSONResponse(builder.get_state().to_dict())
 
 
+@router.get("/testing/structure-instancing")
+async def structure_instancing_example() -> MVSResponse:
+    """
+    Instancing of the same structure
+    """
+    builder = create_builder()
+
+    # instantiate top level structure
+    (
+        builder.download(url=_url_for_mmcif("1tqn"))
+        .parse(format="mmcif")
+        .model_structure()
+        .instance(translation=(-50, 0, 0))
+        .instance(translation=(50, 0, 0))
+        .component()
+        .representation()
+        .color(color="red")
+    )
+
+    # instantiate ligand component
+    (
+        builder.download(url=_url_for_mmcif("1cbs"))
+        .parse(format="mmcif")
+        .model_structure()
+        .component(selector="ligand")
+        .instance(translation=(0, 50, 0))
+        .instance(translation=(0, -50, 0))
+        .representation(type="ball_and_stick")
+        .color(color="blue")
+    )
+
+    return JSONResponse(builder.get_state().to_dict())
+
+
+@router.get("/testing/volume-transform-and-instancing")
+async def volume_transform_and_instancing_example() -> MVSResponse:
+    """
+    Example for transforming and instancing volumes
+    """
+
+    builder = create_builder()
+
+    download = builder.download(url="https://www.ebi.ac.uk/pdbe/entry-files/1cbs.ccp4")
+    volume = download.parse(format="map").volume()
+    (
+        volume.instance(translation=(100, 0, 50))
+        .instance(translation=(-100, 0, -50))
+        .representation(type="isosurface", relative_isovalue=1, show_wireframe=True)
+        .color(color="red")
+    )
+
+    download = builder.download(url="https://www.ebi.ac.uk/pdbe/entry-files/1tqn.ccp4")
+    volume = download.parse(format="map").volume()
+    (
+        volume.transform(translation=(0, -100, 0))
+        .representation(type="isosurface", relative_isovalue=1, show_wireframe=True)
+        .color(color="blue")
+    )
+
+    volume = download.parse(format="map").volume()
+    (volume.representation(type="isosurface", relative_isovalue=1, show_wireframe=True).color(color="green"))
+
+    return JSONResponse(builder.get_state().to_dict())
+
+
 @router.get("/testing/angle-primitive")
 async def testing_angle_primitive_example() -> MVSResponse:
     """
@@ -2324,6 +2420,28 @@ async def testing_angle_primitive_example() -> MVSResponse:
     return JSONResponse(builder.get_state().to_dict())
 
 
+@router.get("/testing/primitives/labels")
+async def testing_primitive_labels_example() -> MVSResponse:
+    """
+    Return a state showing a tethered label attached to a sphere
+    """
+    builder = create_builder()
+    primitives = builder.primitives(
+        label_attachment="top-right",
+        label_show_tether=True,
+        label_tether_length=2,
+        label_background_color="lightblue",
+    )
+    primitives.sphere(
+        center=(10, 10, 10),
+        radius=1,
+        color="red",
+    )
+    primitives.label(position=(10, 10, 10), text="This is a label", label_offset=2)
+
+    return JSONResponse(builder.get_state().to_dict())
+
+
 @router.get("/testing/primitives/from-uri")
 async def primitives_from_uri_example() -> MVSResponse:
     """
@@ -2333,6 +2451,38 @@ async def primitives_from_uri_example() -> MVSResponse:
     builder.primitives_from_uri(uri="http://localhost:9000/api/v1/examples/data/basic-primitives")
 
     return JSONResponse(builder.get_state().to_dict())
+
+
+@router.get("/testing/primitives/snapshot-key")
+async def testing_primitive_snapshot_key_example() -> MVSResponse:
+    """
+    Return a state showing switching states by interacting with primitives
+    """
+
+    def _ping_pong(key: str, snapshot_key: str, attachment: LabelAttachmentT, text: str):
+        builder = create_builder()
+        primitives = builder.primitives(
+            label_attachment=attachment,
+            label_show_tether=True,
+            label_tether_length=2,
+            label_background_color="lightblue",
+            snapshot_key=snapshot_key,
+        )
+        primitives.sphere(
+            center=(10, 10, 10),
+            radius=1,
+            color="red",
+            tooltip="Click to play...",
+        )
+        primitives.label(position=(10, 10, 10), text=text, label_offset=2)
+        return builder.get_snapshot(title=key, key=key)
+
+    snapshots = [
+        _ping_pong("a", "b", "top-right", "Ping"),
+        _ping_pong("b", "a", "bottom-left", "Pong"),
+    ]
+    metadata = GlobalMetadata(description="test")
+    return JSONResponse(States(snapshots=snapshots, metadata=metadata).to_dict())
 
 
 ##############################################################################