+++ CHANGELOG.md
@@ -11,6 +11,12 @@ All notable changes to this project will be documented in this file, following t
 - Add `palette` param for `color_from_*` nodes
 - Add `clip` node support for structure and volume representations
 - Add `instance_id` field to ComponentExpression and MVS annotations
+- Add `grid_slice` volume representation support
+- Add `label_show_tether`, `label_tether_length`, `label_attachment`, and `label_background_color` to `PrimitivesParams`
+- Add `snapshot_key` to `PrimitivesParams` that enables navigating to a different snapshot on interaction
+- Add `matrix` field support to `TransformParams`
+- Add `instance` node type
+- Support transforms and instancing on structures, components, and volumes
 
 ## [v1.6.0] - 2025-04-22